+++ pyproject.toml
@@ -71,6 +71,7 @@ dependencies = [
     "fire",
     "llama_stack_client==0.2.17",
     "pytest-xdist==3.8.0",
+    "curlify>=3.0.0",
 ]
 
 [project.urls]
