+++ tests/conftest.py
@@ -61,6 +61,8 @@ from utilities.operator_utils import get_csv_related_images, get_cluster_service
 
 LOGGER = get_logger(name=__name__)
 
+pytest_plugins = ["tests.fixtures.inference", "tests.fixtures.guardrails", "tests.fixtures.evals"]
+
 
 @pytest.fixture(scope="session")
 def admin_client() -> DynamicClient:
