+++ tests/llama_stack/conftest.py
@@ -1,0 +1,159 @@
+import os
+from typing import Generator, Any, Dict
+
+import portforward
+import pytest
+from _pytest.fixtures import FixtureRequest
+from kubernetes.dynamic import DynamicClient
+from llama_stack_client import LlamaStackClient
+from ocp_resources.data_science_cluster import DataScienceCluster
+from ocp_resources.deployment import Deployment
+from ocp_resources.llama_stack_distribution import LlamaStackDistribution
+from ocp_resources.namespace import Namespace
+from simple_logger.logger import get_logger
+
+from tests.llama_stack.utils import create_llama_stack_distribution, wait_for_llama_stack_client_ready
+from utilities.constants import DscComponents, Timeout
+from utilities.data_science_cluster_utils import update_components_in_dsc
+
+
+LOGGER = get_logger(name=__name__)
+
+
+@pytest.fixture(scope="class")
+def enabled_llama_stack_operator(dsc_resource: DataScienceCluster) -> Generator[DataScienceCluster, Any, Any]:
+    with update_components_in_dsc(
+        dsc=dsc_resource,
+        components={
+            DscComponents.LLAMASTACKOPERATOR: DscComponents.ManagementState.MANAGED,
+        },
+        wait_for_components_state=True,
+    ) as dsc:
+        yield dsc
+
+
+@pytest.fixture(scope="class")
+def llama_stack_server_config(
+    request: FixtureRequest,
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+) -> Dict[str, Any]:
+    fms_orchestrator_url = "http://localhost"
+    inference_model = os.getenv("INFERENCE_MODEL", "")
+    vllm_api_token = os.getenv("VLLM_API_TOKEN", "")
+    vllm_url = os.getenv("VLLM_URL", "")
+
+    if hasattr(request, "param"):
+        if request.param.get("fms_orchestrator_url_fixture"):
+            fms_orchestrator_url = request.getfixturevalue(argname=request.param.get("fms_orchestrator_url_fixture"))
+
+        # Override env vars with request parameters if provided
+        if request.param.get("inference_model"):
+            inference_model = request.param.get("inference_model")
+        if request.param.get("vllm_api_token"):
+            vllm_api_token = request.param.get("vllm_api_token")
+        if request.param.get("vllm_url_fixture"):
+            vllm_url = request.getfixturevalue(argname=request.param.get("vllm_url_fixture"))
+
+    return {
+        "containerSpec": {
+            "resources": {
+                "requests": {"cpu": "250m", "memory": "500Mi"},
+                "limits": {"cpu": "2", "memory": "12Gi"},
+            },
+            "env": [
+                {
+                    "name": "VLLM_URL",
+                    "value": vllm_url,
+                },
+                {"name": "VLLM_API_TOKEN", "value": vllm_api_token},
+                {
+                    "name": "VLLM_TLS_VERIFY",
+                    "value": "false",
+                },
+                {
+                    "name": "INFERENCE_MODEL",
+                    "value": inference_model,
+                },
+                {
+                    "name": "MILVUS_DB_PATH",
+                    "value": "~/.llama/milvus.db",
+                },
+                {"name": "FMS_ORCHESTRATOR_URL", "value": fms_orchestrator_url},
+            ],
+            "name": "llama-stack",
+            "port": 8321,
+        },
+        "distribution": {"image": "quay.io/ruimvieira/llama-stack-lmeval-fix:latest"},
+        "storage": {
+            "size": "20Gi",
+        },
+    }
+
+
+@pytest.fixture(scope="class")
+def llama_stack_distribution(
+    admin_client: DynamicClient,
+    model_namespace: Namespace,
+    enabled_llama_stack_operator: DataScienceCluster,
+    llama_stack_server_config: Dict[str, Any],
+) -> Generator[LlamaStackDistribution, None, None]:
+    with create_llama_stack_distribution(
+        client=admin_client,
+        name="llama-stack-distribution",
+        namespace=model_namespace.name,
+        replicas=1,
+        server=llama_stack_server_config,
+    ) as lls_dist:
+        lls_dist.wait_for_status(status=LlamaStackDistribution.Status.READY, timeout=600)
+        yield lls_dist
+
+
+@pytest.fixture(scope="class")
+def llama_stack_distribution_deployment(
+    admin_client: DynamicClient,
+    llama_stack_distribution: LlamaStackDistribution,
+) -> Generator[Deployment, Any, Any]:
+    deployment = Deployment(
+        client=admin_client,
+        namespace=llama_stack_distribution.namespace,
+        name=llama_stack_distribution.name,
+    )
+
+    deployment.wait(timeout=Timeout.TIMEOUT_2MIN)
+    yield deployment
+
+
+@pytest.fixture(scope="class")
+def llama_stack_client(
+    admin_client: DynamicClient,
+    llama_stack_distribution_deployment: Deployment,
+) -> Generator[LlamaStackClient, Any, Any]:
+    """
+    Returns a ready to use LlamaStackClient,  enabling port forwarding
+    from the llama-stack-server service:8321 to localhost:8321
+
+    Args:
+        admin_client (DynamicClient): Kubernetes dynamic client for cluster operations
+        llama_stack_distribution_deployment (Deployment): LlamaStack distribution deployment resource
+
+    Yields:
+        Generator[LlamaStackClient, Any, Any]: Configured LlamaStackClient for RAG testing
+    """
+    try:
+        with portforward.forward(
+            pod_or_service=f"{llama_stack_distribution_deployment.name}-service",
+            namespace=llama_stack_distribution_deployment.namespace,
+            from_port=8321,
+            to_port=8321,
+            waiting=15,
+        ):
+            client = LlamaStackClient(
+                base_url="http://localhost:8321",
+                timeout=120.0,
+            )
+            wait_for_llama_stack_client_ready(client=client)
+            yield client
+    except Exception as e:
+        LOGGER.error(f"Failed to set up port forwarding: {e}")
+        raise
