+++ tests/model_explainability/guardrails/conftest.py
@@ -1,22 +1,12 @@
-import os
-import subprocess
-from base64 import b64encode
 from typing import Generator, Any
 
 import pytest
-from _pytest.fixtures import FixtureRequest
 from kubernetes.dynamic import DynamicClient
-from ocp_resources.config_map import ConfigMap
-from ocp_resources.deployment import Deployment
-from ocp_resources.guardrails_orchestrator import GuardrailsOrchestrator
 from ocp_resources.inference_service import InferenceService
 from ocp_resources.namespace import Namespace
-from ocp_resources.pod import Pod
-from ocp_resources.resource import ResourceEditor
 from ocp_resources.route import Route
 from ocp_resources.secret import Secret
 from ocp_resources.serving_runtime import ServingRuntime
-from pytest_testconfig import py_config
 
 from utilities.certificates_utils import create_ca_bundle_file
 from utilities.constants import KServeDeploymentType, Labels, RuntimeTemplates, Annotations
@@ -25,125 +15,6 @@ from utilities.serving_runtime import ServingRuntimeFromTemplate
 
 
 GUARDRAILS_ORCHESTRATOR_NAME = "guardrails-orchestrator"
-
-
-# GuardrailsOrchestrator related fixtures
-@pytest.fixture(scope="class")
-def guardrails_orchestrator(
-    request: FixtureRequest,
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    orchestrator_config: ConfigMap,
-) -> Generator[GuardrailsOrchestrator, Any, Any]:
-    gorch_kwargs = {
-        "client": admin_client,
-        "name": GUARDRAILS_ORCHESTRATOR_NAME,
-        "namespace": model_namespace.name,
-        "orchestrator_config": orchestrator_config.name,
-        "replicas": 1,
-        "wait_for_resource": True,
-    }
-
-    if enable_built_in_detectors := request.param.get("enable_built_in_detectors"):
-        gorch_kwargs["enable_built_in_detectors"] = enable_built_in_detectors
-
-    if request.param.get("enable_guardrails_gateway"):
-        guardrails_gateway_config = request.getfixturevalue(argname="guardrails_gateway_config")
-        gorch_kwargs["enable_guardrails_gateway"] = True
-        gorch_kwargs["guardrails_gateway_config"] = guardrails_gateway_config.name
-
-    with GuardrailsOrchestrator(**gorch_kwargs) as gorch:
-        gorch_deployment = Deployment(name=gorch.name, namespace=gorch.namespace, wait_for_resource=True)
-        gorch_deployment.wait_for_replicas()
-        yield gorch
-
-
-@pytest.fixture(scope="class")
-def orchestrator_config(
-    request: FixtureRequest, admin_client: DynamicClient, model_namespace: Namespace
-) -> Generator[ConfigMap, Any, Any]:
-    with ConfigMap(
-        client=admin_client,
-        name="fms-orchestr8-config-nlp",
-        namespace=model_namespace.name,
-        data=request.param["orchestrator_config_data"],
-    ) as cm:
-        yield cm
-
-
-@pytest.fixture(scope="class")
-def guardrails_gateway_config(
-    request: FixtureRequest, admin_client: DynamicClient, model_namespace: Namespace
-) -> Generator[ConfigMap, Any, Any]:
-    with ConfigMap(
-        client=admin_client,
-        name="fms-orchestr8-config-gateway",
-        namespace=model_namespace.name,
-        label={Labels.Openshift.APP: "fmstack-nlp"},
-        data=request.param["guardrails_gateway_config_data"],
-    ) as cm:
-        yield cm
-
-
-@pytest.fixture(scope="class")
-def guardrails_orchestrator_pod(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    guardrails_orchestrator: GuardrailsOrchestrator,
-) -> Pod:
-    return list(
-        Pod.get(
-            namespace=model_namespace.name, label_selector=f"app.kubernetes.io/instance={GUARDRAILS_ORCHESTRATOR_NAME}"
-        )
-    )[0]
-
-
-@pytest.fixture(scope="class")
-def guardrails_orchestrator_route(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    guardrails_orchestrator: GuardrailsOrchestrator,
-) -> Generator[Route, Any, Any]:
-    guardrails_orchestrator_route = Route(
-        name=f"{guardrails_orchestrator.name}",
-        namespace=guardrails_orchestrator.namespace,
-        wait_for_resource=True,
-        ensure_exists=True,
-    )
-    with ResourceEditor(
-        patches={
-            guardrails_orchestrator_route: {
-                "metadata": {
-                    "annotations": {"haproxy.router.openshift.io/timeout": "10m"},
-                }
-            }
-        }
-    ):
-        yield guardrails_orchestrator_route
-
-
-@pytest.fixture(scope="class")
-def guardrails_orchestrator_health_route(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    guardrails_orchestrator: GuardrailsOrchestrator,
-) -> Generator[Route, Any, Any]:
-    guardrails_orchestrator_health_route = Route(
-        name=f"{guardrails_orchestrator.name}-health",
-        namespace=guardrails_orchestrator.namespace,
-        wait_for_resource=True,
-        ensure_exists=True,
-    )
-    with ResourceEditor(
-        patches={
-            guardrails_orchestrator_health_route: {
-                "metadata": {
-                    "annotations": {Annotations.HaproxyRouterOpenshiftIo.TIMEOUT: "10m"},
-                }
-            }
-        }
-    ):
-        yield guardrails_orchestrator_health_route
 
 
 # ServingRuntimes, InferenceServices, and related resources
@@ -214,91 +85,6 @@ def openshift_ca_bundle_file(
     admin_client: DynamicClient,
 ) -> str:
     return create_ca_bundle_file(client=admin_client, ca_type="openshift")
-
-
-@pytest.fixture(scope="class")
-def guardrails_orchestrator_ssl_cert(guardrails_orchestrator_route: Route):
-    hostname = guardrails_orchestrator_route.host
-
-    try:
-        result = subprocess.run(
-            args=["openssl", "s_client", "-showcerts", "-connect", f"{hostname}:443"],
-            input="",
-            capture_output=True,
-            text=True,
-            timeout=30,
-        )
-
-        if result.returncode != 0 and "CONNECTED" not in result.stdout:
-            raise RuntimeError(f"Failed to connect to {hostname}: {result.stderr}")
-
-        cert_lines = []
-        in_cert = False
-        for line in result.stdout.splitlines():
-            if "-----BEGIN CERTIFICATE-----" in line:
-                in_cert = True
-            if in_cert:
-                cert_lines.append(line)
-            if "-----END CERTIFICATE-----" in line:
-                in_cert = False
-
-        if not cert_lines:
-            raise RuntimeError(f"No certificate found in response from {hostname}")
-
-        filepath = os.path.join(py_config["tmp_base_dir"], "gorch_cert.crt")
-        with open(filepath, "w") as f:
-            f.write("\n".join(cert_lines))
-
-        return filepath
-
-    except Exception as e:
-        raise RuntimeError(f"Could not get certificate from {hostname}: {e}")
-
-
-@pytest.fixture(scope="class")
-def guardrails_orchestrator_ssl_cert_secret(
-    admin_client: DynamicClient,
-    model_namespace: Namespace,
-    guardrails_orchestrator_ssl_cert: str,  # ← Add dependency and use correct cert
-) -> Generator[Secret, Any, None]:
-    with open(guardrails_orchestrator_ssl_cert, "r") as f:
-        cert_content = f.read()
-
-    with Secret(
-        client=admin_client,
-        name="orch-certificate",
-        namespace=model_namespace.name,
-        data_dict={"orch-certificate.crt": b64encode(cert_content.encode("utf-8")).decode("utf-8")},
-    ) as secret:
-        yield secret
-
-
-@pytest.fixture(scope="class")
-def patched_llamastack_deployment_tls_certs(llamastack_distribution, guardrails_orchestrator_ssl_cert_secret):
-    lls_deployment = Deployment(
-        name=llamastack_distribution.name,
-        namespace=llamastack_distribution.namespace,
-        ensure_exists=True,
-    )
-
-    current_spec = lls_deployment.instance.spec.template.spec.to_dict()
-
-    current_spec["volumes"].append({
-        "name": "router-ca",
-        "secret": {"secretName": "orch-certificate"},  # pragma: allowlist secret
-    })
-
-    for container in current_spec["containers"]:
-        if container["name"] == "llama-stack":
-            container["volumeMounts"].append({"name": "router-ca", "mountPath": "/etc/llama/certs", "readOnly": True})
-            break
-
-    with ResourceEditor(patches={lls_deployment: {"spec": {"template": {"spec": current_spec}}}}) as _:
-        initial_replicas = lls_deployment.replicas
-        lls_deployment.scale_replicas(replica_count=0)
-        lls_deployment.scale_replicas(replica_count=initial_replicas)
-        lls_deployment.wait_for_replicas()
-        yield lls_deployment
 
 
 @pytest.fixture(scope="class")
