+++ utilities/constants.py
@@ -1,6 +1,8 @@
-from typing import Any
+from typing import Any, Dict
 
 from ocp_resources.resource import Resource
+
+from tests.model_explainability.guardrails.constants import QWEN_ISVC_NAME
 
 
 class KServeDeploymentType:
@@ -355,3 +357,18 @@ OPENSHIFT_OPERATORS: str = "openshift-operators"
 MARIADB: str = "mariadb"
 MODEL_REGISTRY_CUSTOM_NAMESPACE: str = "model-registry-custom-ns"
 THANOS_QUERIER_ADDRESS = "https://thanos-querier.openshift-monitoring.svc:9092"
+BUILTIN_DETECTOR_CONFIG: Dict[str, Any] = {
+    "regex": {
+        "type": "text_contents",
+        "service": {
+            "hostname": "127.0.0.1",
+            "port": 8080,
+        },
+        "chunker_id": "whole_doc_chunker",
+        "default_threshold": 0.5,
+    }
+}
+
+CHAT_GENERATION_CONFIG: Dict[str, Any] = {
+    "service": {"hostname": f"{QWEN_ISVC_NAME}-predictor", "port": 8032, "request_timeout": 600}
+}
